Rename workspace for the generator module to beehive-spirv-toolkit
