+++ generator/pom.xml
@@ -80,7 +80,7 @@
             </goals>
             <configuration>
               <skip>${maven.exec.skip}</skip>
-              <mainClass>uk.ac.manchester.spirvbeehivetoolkit.generator.Runner</mainClass>
+              <mainClass>uk.ac.manchester.beehivespirvtoolkit.generator.Runner</mainClass>
               <arguments>
                 <argument>${project.parent.basedir}/lib/src/main/java/uk/ac/manchester/spirvbeehivetoolkit/lib</argument>
                 <argument>${spirv.gen.majorversion}</argument>
@@ -102,7 +102,7 @@
 
           <transformers>
             <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-              <mainClass>uk.ac.manchester.spirvbeehivetoolkit.generator.Runner</mainClass>
+              <mainClass>uk.ac.manchester.beehivespirvtoolkit.generator.Runner</mainClass>
             </transformer>
           </transformers>
