+++ generator/src/main/java/uk/ac/manchester/beehivespirvtoolkit/generator/MutableInt.java
@@ -23,7 +23,7 @@
  * SOFTWARE.
  */
 
-package uk.ac.manchester.spirvbeehivetoolkit.generator;
+package uk.ac.manchester.beehivespirvtoolkit.generator;
 
 public class MutableInt {
     private int value;
